fixed breaking change made in .net core

+++ X10D.Performant/src/Custom/BooleanExtensions/BooleanExtensions.cs
@@ -6,7 +6,7 @@
     public static class BooleanExtensions
     {
         /// <include file='BooleanExtensions.xml' path='members/member[@name="ToByte"]'/>
-        public static byte ToByte(this bool value) => value ? 1 : 0;
+        public static byte ToByte(this bool value) => (byte)(value ? 1 : 0);
 
         /// <include file='BooleanExtensions.xml' path='members/member[@name="ToDecimal"]'/>
         public static decimal ToDecimal(this bool value) => value ? 1M : 0M;
@@ -15,7 +15,7 @@
         public static double ToDouble(this bool value) => value ? 1D : 0D;
 
         /// <include file='BooleanExtensions.xml' path='members/member[@name="ToInt16"]'/>
-        public static short ToInt16(this bool value) => value ? 1 : 0;
+        public static short ToInt16(this bool value) => (short)(value ? 1 : 0);
 
         /// <include file='BooleanExtensions.xml' path='members/member[@name="ToInt32"]'/>
         public static int ToInt32(this bool value) => value ? 1 : 0;
@@ -27,16 +27,16 @@
         public static nint ToNInt(this bool value) => value ? 1 : 0;
 
         /// <include file='BooleanExtensions.xml' path='members/member[@name="ToUNInt"]'/>
-        public static nuint ToNUInt(this bool value) => value ? 1 : 0;
+        public static nuint ToNUInt(this bool value) => (nuint)(value ? 1 : 0);
 
         /// <include file='BooleanExtensions.xml' path='members/member[@name="ToSByte"]'/>
-        public static sbyte ToSByte(this bool value) => value ? 1 : 0;
+        public static sbyte ToSByte(this bool value) => (sbyte)(value ? 1 : 0);
 
         /// <include file='BooleanExtensions.xml' path='members/member[@name="ToSingle"]'/>
         public static float ToSingle(this bool value) => value ? 1F : 0F;
 
         /// <include file='BooleanExtensions.xml' path='members/member[@name="ToUInt16"]'/>
-        public static ushort ToUInt16(this bool value) => value ? 1 : 0;
+        public static ushort ToUInt16(this bool value) => (ushort)(value ? 1 : 0);
 
         /// <include file='BooleanExtensions.xml' path='members/member[@name="ToUInt32"]'/>
         public static uint ToUInt32(this bool value) => value ? 1U : 0U;
