+++ X10D.Performant/src/Custom/StreamExtensions/WriteExtensions.cs
@@ -9,7 +9,7 @@ namespace X10D.Performant.StreamExtensions
     public static partial class StreamExtensions
     {
         /// <include file='StreamExtensions.xml' path='members/member[@name="WriteBoolean"]'/>
-        public static void Write(this Stream stream, bool value) => stream.WriteByte(value ? 1 : 0);
+        public static void Write(this Stream stream, bool value) => stream.WriteByte((byte)(value ? 1 : 0));
 
         /// <include file='StreamExtensions.xml' path='members/member[@name="WriteDecimal"]'/>
         public static void Write(this Stream stream, decimal value, bool littleEndian = true)
